# spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Contrafactory API
  description: Smart contract artifact registry API
  version: "1.0"
  contact:
    name: API Support
    url: https://github.com/pendergraft/contrafactory/issues
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  termsOfService: https://contrafactory.dev/terms
servers:
  - url: https://api.contrafactory.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
  /api/openapi.yaml:
    get:
      operationId: getOpenApiSpec
      summary: OpenAPI Specification
      description: Download the OpenAPI YAML specification
      tags: []
      security: []
      responses:
        "200":
          description: The OpenAPI specification
        "404":
          description: Specification not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/deployments:
    get:
      operationId: listDeployments
      summary: List deployments
      description: List all deployments with filtering and pagination
      tags: [deployments]
      security: []
      parameters:
        - name: chain_id
          in: query
          description: Filter by chain ID (e.g. 1 for mainnet)
          schema:
            type: string
            example: "1"
        - name: chain
          in: query
          description: Filter by chain name
          schema:
            type: string
            example: mainnet
        - name: package
          in: query
          description: Filter by package name
          schema:
            type: string
        - name: verified
          in: query
          description: Filter by verification status
          schema:
            type: string
            enum: ["true", "false"]
        - name: limit
          in: query
          description: Page limit (max 100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          description: Pagination cursor from previous response
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: recordDeployment
      summary: Record deployment
      description: Record a new contract deployment (requires API key)
      tags: [deployments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordDeploymentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordDeploymentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/packages:
    get:
      operationId: listPackages
      summary: List packages
      description: List all packages with pagination and filtering
      tags: [packages]
      security: []
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: chain
          in: query
          description: Filter by chain name
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field (name, created_at)
          schema:
            type: string
            enum: [name, created_at]
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          description: Page limit (max 100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          description: Pagination cursor
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authenticated endpoints (publish, record, delete)
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Error code (e.g. NOT_FOUND, INVALID_REQUEST)
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          example: Package not found
    Pagination:
      type: object
      required: [limit, hasMore, nextCursor]
      properties:
        limit:
          type: integer
          description: Page size used
        hasMore:
          type: boolean
          description: Whether more results exist
        nextCursor:
          type: string
          description: Cursor for next page (empty if no more)

    # Deployments
    RecordDeploymentRequest:
      type: object
      required: [package, version, contract, chainId, address]
      properties:
        package:
          type: string
          description: Package name
          example: my-contract
        version:
          type: string
          description: Package version
          example: "1.0.0"
        contract:
          type: string
          description: Contract name within the package
          example: MyContract
        chainId:
          type: integer
          description: Chain ID (e.g. 1 for Ethereum mainnet)
          example: 1
        address:
          type: string
          description: Deployed contract address (hex, 0x-prefixed)
          example: "0x1234567890123456789012345678901234567890"
        txHash:
          type: string
          description: Deployment transaction hash
        deployerAddress:
          type: string
          description: Address that deployed the contract
        blockNumber:
          type: integer
          description: Block number of deployment
        constructorArgs:
          type: string
          description: Hex-encoded constructor arguments
        libraries:
          type: object
          additionalProperties:
            type: string
          description: Library address mappings
    RecordDeploymentResponse:
      type: object
      required: [id, chainId, address, verified, message]
      properties:
        id:
          type: string
          description: Deployment record ID
        chainId:
          type: string
          description: Chain ID
        address:
          type: string
          description: Contract address
        verified:
          type: boolean
          description: Whether the deployment is verified
        message:
          type: string
          description: Status message
    DeploymentItem:
      type: object
      properties:
        chainId:
          type: string
        address:
          type: string
        contractName:
          type: string
        verified:
          type: boolean
        txHash:
          type: string
    DeploymentListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    PackageItem:
      type: object
      properties:
        name:
          type: string
        chain:
          type: string
        builder:
          type: string
        versions:
          type: array
          items:
            type: string
          description: List of package versions
    PackageListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PackageItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
...